+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - fixed issue with context files not being highlighted in All files mode on Windows
 - improved parsing of thinking blocks to avoid nested issues
 - using models.dev to load models metadata
+- added more actions for project tabs
 
 ## [0.32.0]